move get_absinfo to raw_stream
